Begin removing internal dependencies on STL
Change data structures to use UVectors or arrays internally,
and include code that dynamically creates a std::vector in header
files rather than .cpp files.

This commit reimplements `SelectorKeys` in this way and replaces the
`VariantMap` with an array of variants (that is turned into a map
dynamically). It adds separate `getKeys()` and `getKeysInternal()`
accessor methods; the first returns a `std::vector`, the second
an array.