Avoid using std::map to represent function options in ExpressionContext; eliminate OrderedMap
Remove internal usage of `std::map` in ExpressionContext.
OrderedMap is no longer necessary since internally, "maps" are represented
as vectors or arrays of pairs. The performance benefits of fast lookup
are probably unnecessary at this point. Ordering is preserved for internal
use; when returning a map, as with function options, preserving the ordering
is unnecessary.