Use std::variant in defining data model types
Use std::variant in definitions of data model classes that
have multiple forms. The previous approach was to (e.g.)
include both a Literal and a Variable member in an Operand
(which may be either a literal or a variable), wasting space.
std::get() is never called and so exceptions should never be thrown.