Change ICU4J JSON parser to follow test schema
Including:
Add code to ICU4J JSON parser to allow `srcs` to be a single element

In a future commit I'll change `srcs` to `src` everywhere and
eliminate `srcs` altogether. For now, `srcs` can be either a single string
or an array of strings. Changed one test to ensure that the parsing code
works.